+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to the package will be documented here.
 - **Updated** *type interfaces* for `toLocaleString` `Chronos` method: **Created** `LocalesArguments` and `DateTimeFormatOptions` *type* and *interface*.
 - **Added** new *public properties* `timeZoneName` and `timeZoneId` to access *current timezone name and identifier(s)*. **Fixed** *UTC offset* related issues.
 - **Adjusted** `native` property to *provide correct native* `Date`. **Fixed** issues with `toDate` and `toLocalISOString`. Optimized internal codes.
+- **Updated** *static* `parse` method to accept *millisecond tokens*.
  -->
 
 ## [4.25.11] - 2025-11-06